# openapi.yaml
openapi: 3.0.3
info:
  title: Fitocube Plant API
  version: 1.0.0
  description: |
    API для платформы Fitocube. Включает регистрацию/логин пользователей (cookie-сессия),
    управление растениями, дружеские запросы, мини-игры и статистику.
tags:
  - name: auth
    description: Регистрация и логин
  - name: plants
    description: Растения
  - name: friends
    description: Друзья
servers:
  - url: http://localhost:8080
    description: Local development
# Authentication & Users
paths:
  /auth/register:
    post:
      summary: Зарегистрировать пользователя по имени
      operationId: registerUser
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
  /auth/login:
    post:
      summary: Выполнить вход по логину и паролю
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход, устанавливается сессионная cookie
          headers:
            Set-Cookie:
              schema:
                type: string
                example: JSESSIONID=abcd1234; HttpOnly; Path=/
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionUser'
        '401':
          description: Неверный логин/пароль
  # Plant Management
  /plants/{plantId}:
    get:
      summary: Получить текущее состояние своего растения
      operationId: getPlantState
      tags: [plants]
      security:
        - SessionCookie: []
      parameters:
        - name: plantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Состояние растения
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantStateDto'
        '404':
          description: Растение не найдено
  /plants/{plantId}/auto-watering:
    post:
      summary: Включить/выключить автополив растения
      operationId: updateAutoWatering
      tags: [plants]
      security:
        - SessionCookie: []
      parameters:
        - name: plantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoWateringRequest'
      responses:
        '200':
          description: Настройки автополива обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantStateDto'
        '400':
          description: Некорректные параметры автополива
        '404':
          description: Растение не найдено
  /plants/by-owner:
    get:
      summary: Получить список растений текущего пользователя
      operationId: listPlantsByOwner
      tags: [plants]
      security:
        - SessionCookie: []
      parameters:
        - name: ownerName
          in: query
          required: false
          deprecated: true
          description: Параметр не используется; список определяется по текущей сессии.
          schema:
            type: string
      responses:
        '200':
          description: Список растений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlantStateDto'
  /plants/by-friend-name:
    get:
      summary: Посмотреть состояние растений друга по его имени (в разработке)
      operationId: listPlantsByFriendName
      tags: [plants]
      security:
        - SessionCookie: []
      parameters:
        - name: friendName
          in: query
          required: true
          schema:
            type: string
      responses:
        '501':
          description: Эндпоинт ещё не реализован
  /plants/watering:
    post:
      summary: Запустить полив растения
      operationId: triggerWatering
      tags: [plants]
      security:
        - SessionCookie: []
      parameters:
        - name: plantId
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WateringRequest'
      responses:
        '202':
          description: Команда на полив отправлена
        '400':
          description: Некорректные параметры полива
        '404':
          description: Растение не найдено
        '503':
          description: MQTT недоступен, команда не отправлена
  /plants/claim:
    post:
      summary: Привязать устройство к текущему пользователю
      operationId: claimPlantDevice
      tags: [plants]
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimRequest'
      responses:
        '201':
          description: Устройство привязано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantStateDto'
        '400':
          description: Некорректный запрос
        '409':
          description: Устройство уже привязано к другому пользователю
  # Friend Management
  /friends:
    get:
      summary: Получить список друзей пользователя
      operationId: listFriends
      tags: [friends]
      security:
        - SessionCookie: []
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список друзей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
  /friends/request:
    post:
      summary: Отправить запрос на добавление в друзья
      operationId: sendFriendRequest
      tags: [friends]
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFriendRequest'
      responses:
        '201':
          description: Запрос создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendRequestDto'
        '404':
          description: Пользователь-цель не найден
  /friends/request/incoming:
    get:
      summary: Посмотреть входящие запросы в друзья
      operationId: listIncomingFriendRequests
      tags: [friends]
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Входящие запросы
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FriendRequestDto'
  /friends/request/outgoing:
    get:
      summary: Посмотреть исходящие запросы в друзья
      operationId: listOutgoingFriendRequests
      tags: [friends]
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Исходящие запросы
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FriendRequestDto'
  /friends/request/{requestId}/accept:
    post:
      summary: Принять запрос в друзья
      operationId: acceptFriendRequest
      tags: [friends]
      security:
        - SessionCookie: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Запрос принят
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: Запрос не найден
  /friends/request/{requestId}/reject:
    post:
      summary: Отклонить запрос в друзья
      operationId: rejectFriendRequest
      tags: [friends]
      security:
        - SessionCookie: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Запрос отклонён
        '404':
          description: Запрос не найден
components:
  schemas:
    SessionUser:
      type: object
      properties:
        id:
          type: integer
        userName:
          type: string
        displayName:
          type: string
    UserDto:
      type: object
      properties:
        userId:
          type: integer
        userName:
          type: string
        displayName:
          type: string
          description: Дополнительное имя для отображения
        passwordHash:
          type: string
          description: Пароль пользователя
        friends:
          type: array
          description: Список идентификаторов друзей пользователя
          items:
            type: integer
    CreateUserRequest:
      type: object
      required:
        - userName
        - password
      properties:
        userName:
          type: string
        displayName:
          type: string
        password:
          type: string
          format: password
          description: Пароль для входа
    LoginRequest:
      type: object
      required:
        - userName
        - password
      properties:
        userName:
          type: string
        password:
          type: string
          format: password
    ClaimRequest:
      type: object
      required:
        - deviceUid
      properties:
        deviceUid:
          type: string
          description: Уникальный код устройства
        nickname:
          type: string
          description: Имя растения
    WateringRequest:
      type: object
      properties:
        durationSeconds:
          type: integer
          minimum: 1
          maximum: 600
          description: Длительность полива в секундах (опционально)
    AutoWateringRequest:
      type: object
      required:
        - enabled
      properties:
        enabled:
          type: boolean
        thresholdPercent:
          type: integer
          minimum: 0
          maximum: 100
          description: Порог влажности почвы для автополива (опционально)
        durationSeconds:
          type: integer
          minimum: 1
          maximum: 600
          description: Длительность включения помпы (опционально)
        cooldownSeconds:
          type: integer
          minimum: 0
          maximum: 86400
          description: Минимальный интервал между поливами (опционально)
    PlantStateDto:
      type: object
      properties:
        plantId:
          type: integer
        deviceUid:
          type: string
          description: Аппаратный идентификатор устройства
        owner:
          $ref: '#/components/schemas/UserDto'
        nickname:
          type: string
          description: Название растения, заданное владельцем
        measurements:
          type: object
          properties:
            airTemperatureC:
              type: number
              format: float
            airHumidityPercent:
              type: number
              format: float
            soilMoisturePercent:
              type: number
              format: float
            soilMoistureRaw:
              type: integer
            timestamp:
              type: string
              format: date-time
        mood:
          type: string
          enum:
            - happy
            - normal
            - thirsty
            - dry
            - hot
            - cold
        friendVisible:
          type: boolean
        autoWateringEnabled:
          type: boolean
        autoWateringThresholdPercent:
          type: integer
          minimum: 0
          maximum: 100
        autoWateringDurationSeconds:
          type: integer
          minimum: 1
          maximum: 600
        autoWateringCooldownSeconds:
          type: integer
          minimum: 0
          maximum: 86400
        lastWateringAt:
          type: string
          format: date-time
      example:
        plantId: 42
        deviceUid: "esp32-d34-202501"
        owner:
          userId: 7
          userName: "montana"
          displayName: "Nikola"
          friends: [3, 9]
        nickname: "Фикус"
        measurements:
          airTemperatureC: 23.5
          airHumidityPercent: 48.0
          soilMoisturePercent: 35.2
          timestamp: "2025-01-25T12:00:00Z"
        mood: happy
        friendVisible: true
        autoWateringEnabled: false
    CreateFriendRequest:
      type: object
      required:
        - requesterId
        - targetName
      properties:
        requesterId:
          type: integer
        targetName:
          type: string
    FriendRequestDto:
      type: object
      properties:
        requestId:
          type: integer
        requester:
          $ref: '#/components/schemas/UserDto'
        target:
          $ref: '#/components/schemas/UserDto'
        status:
          type: string
          enum:
            - pending
            - accepted
            - rejected
        createdAt:
          type: string
          format: date-time
    PlayerStatsDto:
      type: object
      properties:
        userId:
          type: integer
        totalHappyHours:
          type: number
          format: float
        moodBattlesWon:
          type: integer
        moodBattlesLost:
          type: integer
        dropsEarned:
          type: integer
        lastUpdated:
          type: string
          format: date-time
    CreateMoodBattleRequest:
      type: object
      required:
        - challengerPlantId
        - opponentPlantId
      properties:
        challengerPlantId:
          type: integer
        opponentPlantId:
          type: integer
        durationHours:
          type: integer
          description: Длительность батла (по умолчанию 24 часа)
    MoodBattleDto:
      type: object
      properties:
        battleId:
          type: integer
        challengerPlantId:
          type: integer
        opponentPlantId:
          type: integer
        challengerScoreHours:
          type: number
          format: float
          description: Сколько часов растение было в состоянии happy
        opponentScoreHours:
          type: number
          format: float
        status:
          type: string
          enum: [pending, active, finished]
        startedAt:
          type: string
          format: date-time
        finishesAt:
          type: string
          format: date-time
        winnerPlantId:
          type: integer
          nullable: true
        dropsReward:
          type: integer
          description: Сколько капель начисляется победителю
    PlantChallengeDto:
      type: object
      properties:
        challengeId:
          type: integer
        plantId:
          type: integer
        weekStart:
          type: string
          format: date
        goalHappyHours:
          type: number
          format: float
        achievedHappyHours:
          type: number
          format: float
        status:
          type: string
          enum: [in_progress, completed, failed]
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Сессионная cookie (HttpOnly). Получается после /auth/login и используется для всех защищённых запросов.